Add the learning engine's observation, classification and derivation layers, and all 36 rules
Learning engine
- observe.py: typed, scoped, weighted observations plus the shared traversal
  every deriver uses, so colours are weighted by the area they cover and fonts
  by the characters they set rather than by how many shapes carry them
- classify.py: the invariant/dominant/multimodal/variable cascade from section
  8.2, with entropy as the guard against mistaking noise for a set of
  conventions, plus the Derivation accumulator that records provenance at the
  moment of derivation rather than reconstructing it later
- derive_brand.py: palette clustered in Lab space with the discarded tail
  reported rather than admitted, a tolerance derived from half the minimum
  inter-cluster distance, typeface and per-role size derivation, logo geometry
  per archetype with absent archetypes marked exempt, page-number format and
  position, and boilerplate
- derive_layout.py: slide size, safe margins, the column and row grid, and
  recurring element positions, resolving a multimodal key by archetype so "the
  title has two positions" becomes two usable rules
- derive_typography.py: every convention, each counted at the scope that makes
  it meaningful, delegating all parsing to tieout/text.py so a derived
  convention and its enforcement cannot disagree
- derive_terms.py: canon candidates, which are proposed rather than decided,
  and are only turned into questions when the variation is not explained by
  sentence-initial capitalisation

Rules: all 36 implemented, registered, documented and tested. Verified against
the generated fixtures: the clean deck produces zero findings from every
category, each rule catches its own seeded defect, and each single-defect
variant triggers only its own rule.

Shared primitives: cluster.py and text.py as described in the previous commit,
now exercised by both the derivers and the rules.

Bugs found and fixed while integrating, each of which would have produced
findings on a correct deck:

- _source_part_for_rels mapped the package's root _rels/.rels to "/", so every
  relationship target resolved with a leading slash and every well-formed deck
  reported four broken relationships
- the furniture cache was keyed on id(), which CPython recycles, so a
  short-lived profile could silently inherit a collected profile's furniture;
  it is now weak-keyed on the deck with an identity check on the profile
- DeckModel carried the default dataclass __eq__, which suppressed __hash__ and
  made it unusable as a cache key; it now has identity semantics, which is also
  the correct model for a mutable aggregate
- capitalised_ngrams joined runs across punctuation, so "Source: Company
  management" produced the phantom term "Source Company", and every footnote
  beginning "Source:" pushed a spurious terminology question
- canon candidates were proposed for ordinary capitalised English, which would
  have spent the interview's twelve-question budget before reaching anything the
  user needed to decide
- the fixture declared a fifth palette colour no element used, and the HY-008
  seed replaced the logo with a low-resolution re-export whose different SHA1
  made BR-001 report a missing logo on the same slide
- the fixture had only two dates, below min_support, so the date convention was
  not learnable and TY-008 could not be tested

ruff check and mypy --strict are clean. RUF001-003 are disabled with a stated
rationale: detecting curly quotes, en dashes and non-breaking spaces is what
this tool does, so those characters must appear literally.

+++ tieout/model/furniture.py
@@ -187,7 +187,7 @@ def _logo_hashes(
         declared = tuple(profile.brand.logo.image_sha1)
         return declared, {sha: support.get(sha, 0) for sha in declared}
 
-    threshold = max(1, int(round(LOGO_SUPPORT_SHARE * deck.slide_count)))
+    threshold = max(1, round(LOGO_SUPPORT_SHARE * deck.slide_count))
     candidates = [sha for sha, count in support.items() if count >= threshold]
     candidates.sort(key=lambda sha: (-support[sha], sha))
     return tuple(candidates), support
@@ -283,7 +283,7 @@ def _boilerplate(
                 continue
             counts.setdefault(normalise_text(text), set()).add(slide.index)
 
-    threshold = max(2, int(round(BOILERPLATE_SUPPORT_SHARE * deck.slide_count)))
+    threshold = max(2, round(BOILERPLATE_SUPPORT_SHARE * deck.slide_count))
     return {
         key: tuple(sorted(slides))
         for key, slides in counts.items()
